Moved showcase to the main module
