+++ WooCommerce/src/debug/AndroidManifest.xml
@@ -37,6 +37,11 @@
             android:authorities="${applicationId}.leakcanary-installer"
             android:exported="false" />
 
+        <activity
+            android:theme="@style/Theme.AppCompat.DayNight.NoActionBar"
+            android:name="com.woocommerce.android.iapshowcase.IAPShowcaseActivity"
+            android:exported="true" />
+
     </application>
 
 </manifest>
