--- iapshowcase/build.gradle
@@ -1,64 +1,0 @@
-plugins {
-    id 'com.android.application'
-    id 'org.jetbrains.kotlin.android'
-}
-
-android {
-    compileSdkVersion gradle.ext.compileSdkVersion
-
-    defaultConfig {
-        minSdkVersion gradle.ext.minSdkVersion
-        targetSdkVersion gradle.ext.targetSdkVersion
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-}
-
-repositories {
-    maven {
-        url 'https://a8c-libs.s3.amazonaws.com/android'
-        content {
-            includeGroup "org.wordpress"
-            includeGroup "org.wordpress.fluxc"
-            includeGroup "org.wordpress.fluxc.plugins"
-            includeGroup "org.wordpress.wellsql"
-            includeGroup "org.wordpress.mediapicker"
-            includeGroup "com.automattic"
-            includeGroup "com.automattic.tracks"
-        }
-    }
-    maven {
-        url "https://a8c-libs.s3.amazonaws.com/android/jcenter-mirror"
-        content {
-            includeVersion "com.android.volley", "volley", "1.1.1"
-            includeVersion "com.google.android", "flexbox", "2.0.1"
-            includeVersion "com.jraska", "falcon", "2.1.1"
-            includeVersion "org.wordpress", "wellsql", "1.6.0"
-            includeVersion "org.wordpress", "wellsql-core", "1.6.0"
-        }
-    }
-}
-
-dependencies {
-    implementation project(":libs:iap")
-    implementation("${gradle.ext.fluxCBinaryPath}:$fluxCVersion") {
-        exclude group: "com.android.support"
-        exclude group: "org.wordpress", module: "utils"
-    }
-    implementation("${gradle.ext.fluxCWooCommercePluginBinaryPath}:$fluxCVersion") {
-        exclude group: "com.android.support"
-        exclude group: "org.wordpress", module: "utils"
-    }
-
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutinesVersion"
-
-    implementation("com.android.billingclient:billing-ktx:$billingVersion")
-    implementation "androidx.appcompat:appcompat:$appCompatVersion"
-
-    implementation "androidx.activity:activity-ktx:1.5.1"
-    implementation "androidx.lifecycle:lifecycle-viewmodel:$lifecycleVersion"
-}
