--- iapshowcase/src/main/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.woocommerce.android.iapshowcase">
-
-    <application>
-        <activity
-            android:theme="@style/Theme.AppCompat.DayNight.NoActionBar"
-            android:name="com.woocommerce.android.iapshowcase.IAPShowcaseActivity"
-            android:exported="true" />
-    </application>
-</manifest>
