+++ libs/iap/src/main/java/com/woocommerce/android/iap/internal/core/IAPManager.kt
@@ -12,13 +12,13 @@ import com.android.billingclient.api.queryPurchasesAsync
 import com.woocommerce.android.iap.internal.model.IAPProductDetailsResponse
 import com.woocommerce.android.iap.internal.model.IAPProductDetailsResponse.Error
 import com.woocommerce.android.iap.internal.model.IAPProductDetailsResponse.Success
-import com.woocommerce.android.iap.public.IAP_LOG_TAG
-import com.woocommerce.android.iap.public.LogWrapper
-import com.woocommerce.android.iap.public.model.IAPBillingErrorType
-import com.woocommerce.android.iap.public.model.IAPProduct
-import com.woocommerce.android.iap.public.model.IAPProductInfoResponse
-import com.woocommerce.android.iap.public.model.IAPProductType
-import com.woocommerce.android.iap.public.model.IAPPurchaseResponse
+import com.woocommerce.android.iap.pub.IAP_LOG_TAG
+import com.woocommerce.android.iap.pub.LogWrapper
+import com.woocommerce.android.iap.pub.model.IAPBillingErrorType
+import com.woocommerce.android.iap.pub.model.IAPProduct
+import com.woocommerce.android.iap.pub.model.IAPProductInfoResponse
+import com.woocommerce.android.iap.pub.model.IAPProductType
+import com.woocommerce.android.iap.pub.model.IAPPurchaseResponse
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlin.coroutines.suspendCoroutine
