+++ libs/iap/src/main/java/com/woocommerce/android/iap/internal/model/IAPProductDetailsResponse.kt
@@ -1,7 +1,7 @@
 package com.woocommerce.android.iap.internal.model
 
 import com.android.billingclient.api.ProductDetails
-import com.woocommerce.android.iap.public.model.IAPBillingErrorType
+import com.woocommerce.android.iap.pub.model.IAPBillingErrorType
 
 internal sealed class IAPProductDetailsResponse {
     data class Success(val productDetails: List<ProductDetails>) : IAPProductDetailsResponse()
