+++ libs/iap/src/main/java/com/woocommerce/android/iap/pub/IAPLogWrapper.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.iap.public
+package com.woocommerce.android.iap.pub
 
 const val IAP_LOG_TAG = "IAP"
