+++ libs/iap/src/main/java/com/woocommerce/android/iap/pub/IAPStore.kt
@@ -1,0 +1,7 @@
+package com.woocommerce.android.iap.pub
+
+import com.woocommerce.android.iap.pub.model.IAPProduct
+
+interface IAPStore {
+    fun confirmOrderOnServer(iapProduct: IAPProduct)
+}
