--- libs/iap/src/main/java/com/woocommerce/android/iap/public/IAPStore.kt
@@ -1,7 +1,0 @@
-package com.woocommerce.android.iap.public
-
-import com.woocommerce.android.iap.public.model.IAPProduct
-
-interface IAPStore {
-    fun confirmOrderOnServer(iapProduct: IAPProduct)
-}
